requirements.txt
notion-client==2.2.1
dropbox==12.0.2
playwright==1.48.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
aiohttp==3.10.5
